Faster discovery by not loading/unloading keyboard
The button with the search glass is actually a text input;
pressing it will place a cursor and open a keyboard. This keyboard is then hidden rightaway.
Loading/unloading the keyboard takes a long time especially on Android phones,
which makes the already slow process of loading the discovery list even slower.

By marking the inner input field as readOnly, the OS knows you don't need a keyboard and won't show one.