Improve handling of Terms of Use
Remove old code that checked if the ToU contains a link; earlier
iterations of the app assumed that the ToU was ONLY a link, but that's
not how it's supposed to work.

Update the viewer to support UTF-8, only ASCII was supported earlier.

Handle newlines in the ToU text by replacing single newlines with <br>
and multiple newlines with <p>.